+++ src/count_collisions.h
@@ -1,8 +1,19 @@
-#ifndef THESIS_WORK_TESTS_H
-#define THESIS_WORK_TESTS_H
+#ifndef THESIS_WORK_COUNT_COLLISIONS_H
+#define THESIS_WORK_COUNT_COLLISIONS_H
 
+#include <iostream>
+#include <map>
+#include <set>
+#include <string_view>
 #include <unordered_map>
 
-size_t CountCollisions(const std::unordered_map<size_t, size_t>& hashes);
+struct CollisionsInfo {
+    size_t count;
+    std::set<std::string_view> words;
+};
 
-#endif //THESIS_WORK_TESTS_H
+using CollisionsMap = std::map<size_t, CollisionsInfo>;
+
+size_t CountCollisions(const CollisionsMap& hashes, std::ostream& out);
+
+#endif //THESIS_WORK_COUNT_COLLISIONS_H