+++ src/img.cpp
@@ -26,8 +26,6 @@ namespace img {
         return height_;
     }
 
-    // шаг задаёт смещение соседних строк изображения
-    // он обычно совпадает с width, но может быть больше
     int Image::GetStep() const {
         return step_;
     }